\name{byumpa-package}
\alias{byumpa-package}
\alias{byumpa}
\docType{package}
\title{
    BYU MPA functions
}
\description{
    More about what it does
}
\details{
\tabular{ll}{
Package: \tab byumpa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-11-24\cr
License: \tab MIT\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{

}
\examples{

}
